Fixed a null reference bug
C# initializes arrays of custom objects to null. There is no need for
this method call here since ParticleMovement[1] is already null. Should
there be alternative error checking? At this point, ParticleMovement[1]
is only ever used and initialized is there is indeed a second particle.
Otherwise, it is never used.